Create Abstract-vs-Summary
This program uses ArXiv API to directly pull the summary for an article.  Then Falcon AI API is used to summarize the article.  Additionally, the article was trimmed so that only the body of the paper will be summarized.  This avoids additional summaries of cited sources, headers, etc that are not wanted.   I am still fine tuning the settings on length of summaries, and brainstorming on how to make the summaries more cohesive.  I am again using the method of splitting the paper into chunks to summarize because it is beyond the context window limits. This does technically work but it does not return cohesive answers sometimes.  There are instances of words that drop off, incomplete sentences, etc. that need to be cleaned up.