Create Semantic-Scholar
This is my first attempt using Semantic Scholar to retrieve information on citations.  I noticed that depending on how new the paper is, Semantic Scholar might not have it in its database yet.  Also, I chose a popular paper and it retrieved 1000 citations but nearly all of them are empty of information besides title.  The authors, abstract, etc are mostly all blank.  I will continue to use Semantic Scholar and see if it is worth using as a tool in this project.